% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{add_explicit_file_names_to_config}
\alias{add_explicit_file_names_to_config}
\title{Add File Names to Existing Config}
\usage{
add_explicit_file_names_to_config(config_path, cholera_directory)
}
\arguments{
\item{config_path}{File name of config}

\item{cholera_directory}{Path to cholera-mapping-pipeline checkout}
}
\description{
Take a config, read it, determine it's default filenames, and add them explicitly to it's config.  WARNING : This function will overwrite the config in question
}
