% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{aggregate_raster_xcells}
\alias{aggregate_raster_xcells}
\title{aggregate_raster_xcells}
\usage{
aggregate_raster_xcells(x, fun, ...)
}
\arguments{
\item{x}{Raster* object}

\item{fun}{function that returns a single value, e.g. mean or min, and that takes a na.rm argument (or can pass through arguments via ...)}

\item{...}{additional arguments as for writeRaster}
}
\value{
A vector
}
\description{
Apply a function to each layer, aggregating the cells of the raster
}
