% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{aggregate_raster_xlayers}
\alias{aggregate_raster_xlayers}
\title{aggregate_raster_xlayers}
\usage{
aggregate_raster_xlayers(x, fun, ...)
}
\arguments{
\item{x}{Raster* object}

\item{fun}{function that returns a single value, e.g. mean or min, and that takes a na.rm argument (or can pass through arguments via ...)}

\item{...}{additional arguments as for writeRaster}
}
\value{
A RasterLayer
}
\description{
A mask of stackApply that always assumes indices is all 1
}
