% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{aggregate_taxonomy_data}
\alias{aggregate_taxonomy_data}
\title{aggregate_taxonomy_data}
\usage{
aggregate_taxonomy_data(
  data,
  ISO_level = 2,
  temporal_aggregate_time_unit = "year",
  aggregate_columns = "sCh",
  observation = TRUE,
  min_total_length = 60,
  max_overlap = 8,
  filter_NA_cases = TRUE,
  time_combine = "none"
)
}
\arguments{
\item{data}{a \code{tbl_dt(data.frame} containing the data.)}

\item{ISO_level}{spatial level to aggreate_to.  Either 0 for country level, a positive integer for ISO_A2_L?, 'official' to the nearest official shapefile, or Inf, for no aggregation}

\item{temporal_aggregate_time_unit}{Currently does nothing.  Intended to allow for considering observations at multiple time aggregates (or none)}

\item{aggregate_columns}{which columns to aggregate over.}

\item{min_total_length}{In order to shunt cases into one year instead of dividing them, how many days does an observation need to cover?}

\item{max_overlap}{In order to shunt cases into one year instead of dividing them, how many days can be removed from a year.}

\item{filter_NA_cases}{Whether or not to remove cases which are NA (as opposed to 0)}

\item{time_combine}{If 'strict', require time periods to match end to end to make an observation.  Otherwise, just assume all grouped observations are the same.}
}
\value{
A \code{tbl_dt(data.frame} the aggregated data.)
}
\description{
This function groups data by certain fields and sums over places where those fields match.
}
