% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{align_rasters2}
\alias{align_rasters2}
\title{align_rasters2}
\usage{
align_rasters2(
  unaligned,
  reference,
  dstfile,
  output_raster = FALSE,
  nThreads = 1,
  projres_only = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Modified version of gdalUtils::align_rasters to allow for unquoted dataset names, which enables the use of postgis layers, the rest of the codes were untouched
}
