% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{apply_to_all_sublevels}
\alias{apply_to_all_sublevels}
\title{apply_to_all_sublevels}
\usage{
apply_to_all_sublevels(
  location_name,
  raster_layer,
  ISO_level,
  fun,
  taxonomy_dir = "taxonomy-verified",
  verbose = FALSE,
  fix_location_names = FALSE,
  partial_cover = FALSE,
  trim_small = 0,
  method = "not raster",
  ...
)
}
\arguments{
\item{location_name}{The name of the country}

\item{ISO_level}{The level to get the shapefiles at}

\item{taxonomy_dir}{The taxonomy directory to store the shapefiles in}

\item{verbose}{Set to true for more warnings and messages}

\item{partial_cover}{Set to true to get accurate partial matches by subsampling the raster grid.}

\item{raster}{The raster to apply the function to}
}
\description{
Apply a function to a raster with different answers for each subregion of a country at a particular ISO level
}
