% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_helpers.R
\name{array_to_adjacency}
\alias{array_to_adjacency}
\title{array_to_adjacency}
\usage{
array_to_adjacency(arr, sparse = TRUE)
}
\arguments{
\item{arr}{array The array to get the centers of}

\item{sparse}{boolean whether to return a sparse or dense matrix}
}
\description{
get the adjacency matrix of the gridcells of an array
}
