% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_helpers.R
\name{array_to_centers}
\alias{array_to_centers}
\title{array_to_centers}
\usage{
array_to_centers(arr)
}
\arguments{
\item{arr}{array The array to get the centers of}
}
\description{
get the centroids of the gridcells of an array
}
