% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{automate_mapping_config}
\alias{automate_mapping_config}
\title{Automate generation of config files}
\usage{
automate_mapping_config(p, covariate_names, countries_names, countries_ids)
}
\arguments{
\item{p}{dataframe with config parameters, one config per row}

\item{covariate_names}{character vector with names of covariates to include in the model}

\item{countries_names}{character vector with country codes to include in the model}

\item{countries_ids}{character vector with country location ids to include in the model}
}
\value{
string with yaml config file
}
\description{
Automate generation of model configuration files from a dataframe of parameters. Generates one config per row in the p dataframe.
}
