% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{build_geoms_query}
\alias{build_geoms_query}
\title{Build geometry}
\usage{
build_geoms_query(conn, schema = "public", table_name, type = "centroids")
}
\arguments{
\item{conn}{connection to the database, object of type DBI::dbConnect}

\item{schema}{schema of the database where the raster is stored}

\item{table_name}{name of the raster to process}

\item{type}{type of geometry to compute, needs to be either 'centroids' or
'polygons'}
}
\value{
None
}
\description{
Builds the centroids or the polygons corresponding to a given raster and writes the result to a new table in the database
}
