% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_helpers.R
\name{case_definition_to_column_name}
\alias{case_definition_to_column_name}
\title{case_definition_to_column_name}
\usage{
case_definition_to_column_name(type, database = FALSE, sql = FALSE)
}
\arguments{
\item{type}{string of type}

\item{database}{Whether or not we're using the database}

\item{sql}{logical for whether data is pulled from sql}
}
\value{
string of column names in the data taxonomy data frame.
}
\description{
Turns human readable types of cholera case definitions into taxdat codes
}
