% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_case_definition}
\alias{check_case_definition}
\title{Check case definition}
\usage{
check_case_definition(case_col)
}
\arguments{
\item{case_column}{the name of the case column to use in the database}
}
\value{
case column if valid
}
\description{
Checks whether the column name for case definition is valid
}
