% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_covariate_choices}
\alias{check_covariate_choices}
\title{Check covariate choices}
\usage{
check_covariate_choices(covar_choices, available_choices)
}
\arguments{
\item{covar_choices}{vector of user-defined covariate choices}

\item{available_choices}{vector of available covaraites choices}
}
\value{
if valid the vector of covariate choices
}
\description{
Verifies whether user-defined covariate choices are in the available
set
}
