% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_model_date_range}
\alias{check_model_date_range}
\title{Check modeling date range}
\usage{
check_model_date_range(
  start_time,
  end_time,
  time_change_func,
  aggregate_to_start,
  aggregate_to_end
)
}
\arguments{
\item{start_time}{the start time of the modeling time rante}

\item{end_time}{the end time of the modeling time range}

\item{time_change_func}{function to change dates to the aggregation level
determined by the time resolution of the model}

\item{aggregate_to_start}{function to aggregate dates to the start date of modeling time slices}

\item{aggregate_to_end}{function to aggregate dates to the end of the modeling time slices}
}
\value{
return
}
\description{
Checks if input date ranges are correct
}
