% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_set_tfrac}
\alias{check_set_tfrac}
\title{Check set tfrac}
\usage{
check_set_tfrac(set_tfrac)
}
\arguments{
\item{config}{the config}
}
\value{
if valid the stan model path
}
\description{
Checks whether the user-specified value of tfrac is valid
}
