% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_stan_model}
\alias{check_stan_model}
\title{Check stan model}
\usage{
check_stan_model(stan_model_path, stan_dir)
}
\arguments{
\item{stan_model_path}{the path to the stan model to use}

\item{stan_dir}{the directory to the directory with all available stan models}
}
\value{
if valid the stan model path
}
\description{
Checks whether the stan model file exists
}
