% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{check_time_res}
\alias{check_time_res}
\title{Check time resolution}
\usage{
check_time_res(res_time)
}
\arguments{
\item{res_time}{the time resolution of the model}
}
\value{
res_time if valid
}
\description{
Checks whether the time resolution input is valid
}
