% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{classify_date}
\alias{classify_date}
\title{classify_date}
\usage{
classify_date(date)
}
\arguments{
\item{date}{The length of time as a difftime.}
}
\description{
Take the length of a report and classify it as daily, weekly, biweekly, monthly, bimonthly, yearly, or biyearly.
}
