% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{clean_all_tmp}
\alias{clean_all_tmp}
\title{clean all tmp}
\usage{
clean_all_tmp(dbuser, map_name)
}
\arguments{
\item{dbuser}{database username}

\item{map_name}{A string representing a somewhat unique name for this run}
}
\description{
Delete all temporary tables in the database used to create this map
}
