% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scale.R
\name{color_scale}
\alias{color_scale}
\title{color_scale}
\usage{
color_scale(type = "cases", use_case = "leaflet", use_log = FALSE)
}
\arguments{
\item{type}{string indicating 'cases' or 'rates' (default: 'cases')}

\item{use_case}{string indicating type of map 'leaflet' or 'ggplot map' (default: 'leaflet')}

\item{use_log}{logical for whether outcome should be log10 transformed (default: FALSE)}
}
\value{
leaflet or ggplot color palette for mapping
}
\description{
default color scale for cholera maps
}
