% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to database}
\usage{
connect_to_db(dbuser)
}
\arguments{
\item{dbuser}{database user name}
}
\value{
a DBI database connection object

db connection object
}
\description{
Helper to connect to cholera_covariates database

Connects to the postgres/postgis cholera_covariates database
}
