% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{country_data_overview}
\alias{country_data_overview}
\title{country_data_overview}
\usage{
country_data_overview(
  country,
  taxonomy.dir = "taxonomy-verified",
  simplified = TRUE,
  pretty = TRUE
)
}
\arguments{
\item{country}{name of country}

\item{taxonomy.dir}{taxonomy directory}

\item{simplified}{if TRUE returns a simplified overview of DESC files, if FALSE a table with all details from DESC files}
}
\value{
knitr table with country data overview
}
\description{
produces an overview of data for a country
}
