% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{create_distance_to_shapefile_raster}
\alias{create_distance_to_shapefile_raster}
\title{create_distance_to_shapefile_raster}
\usage{
create_distance_to_shapefile_raster(
  shapefile,
  region = NULL,
  layers_dir = "Layers",
  resol = 10
)
}
\arguments{
\item{shapefile}{The shapefile of features}

\item{region}{A worldpop region containing the shapefile (or NULL to lookup automatically)}

\item{resol}{Approximate resolution of the output raster in km}
}
\value{
A RasterLayer
}
\description{
Create a RasterLayer that is 0 on a shapefile and the distance to the nearest cell intersecting the shapefile otherwise
}
