% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_progress_functions.R
\name{create_mapping_progress_table}
\alias{create_mapping_progress_table}
\title{create_mapping_progress_table}
\usage{
create_mapping_progress_table(config_path, pipeline_path)
}
\arguments{
\item{config_path}{A filename or directory containing configs which should be used.  This function accepts multiple config_paths in this argument, and the results are concatenated by row.}
}
\value{
a data.frame with rows based on configs, and columns based on different stages of progress, with TRUE if the stage has been completed for that config, and FALSE if it has not
}
\description{
From a list of configs (or directories containing only configs), create a table displaying the files which exist for those configs, detailing which are present and which are absent
}
