% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{create_multiple_test_covariates}
\alias{create_multiple_test_covariates}
\title{create multiple test covariates}
\usage{
create_multiple_test_covariates(
  test_raster = create_test_raster(),
  ncovariates = 5,
  nonspatial = c(TRUE, FALSE, FALSE, FALSE, FALSE),
  nontemporal = c(FALSE, TRUE, FALSE, FALSE, FALSE, TRUE),
  spatially_smooth = c(FALSE, FALSE, TRUE, TRUE, FALSE),
  temporally_smooth = c(FALSE, FALSE, FALSE, TRUE, FALSE, TRUE),
  polygonal = c(TRUE, FALSE, FALSE, FALSE, FALSE),
  radiating = c(FALSE, TRUE, FALSE, FALSE, TRUE),
  constant = c(FALSE, FALSE, FALSE, FALSE, FALSE),
  rho = rep(0.999999, times = ncovariates),
  polygons = create_test_layered_polygons(),
  radiating_polygons = list(NA, create_test_polygons(dimension = 0, number = 2), NA,
    NA, sf::st_union(create_test_polygons(dimension = 1))),
  smoothing_function = rep(list(function(n, mu, covariance, centers) {    
    return(scale(MASS::mvrnorm(n = n, mu = mu, Matrix::solve(covariance)))) }),
    ncovariates),
  radiation_function = rep(list(function(x, mu) {     mu * exp(-(x/10000)^2) }),
    ncovariates),
  radiating_means = list(NA, rnorm(2), NA, NA, 1),
  weights = rep(list(c(0.3, 1, 1, 1, 1)), ncovariates),
  magnitude = rnorm(ncovariates),
  family = "Gaussian",
  seed
)
}
\arguments{
\item{test_raster}{raster::raster The raster to make the covariate over.
It should have one layer per time step}

\item{ncovariates}{integer the number of covariates to make}

\item{nonspatial}{boolean(ncovariates) Whether the independent layer varies in space for each covariate}

\item{nontemporal}{boolean(ncovariates) Whether the independent layer varies in time for each covariate}

\item{spatially_smooth}{boolean(ncovariates) Whether the locally correlated component is spatially correlated for each covariate}

\item{temporally_smooth}{boolean(ncovariates) Whether the locally correlated component is temporally correlated for each covariate}

\item{polygonal}{boolean(ncovariates) Whether to include the polygonal layer for each covariate}

\item{radiating}{boolean(ncovariates) Whether to include the radiating layer for each covariate}

\item{constant}{boolean(ncovariates) Whether to include a constant layer for each covariate}

\item{rho}{numeric(ncovariates) A parameter governing spatial correlation used to determine the covariance matrix for each covariate}

\item{polygons}{sf::sf The polygons to use for the polygonal layer}

\item{radiating_polygons}{list(sf::sf) The polygons that the radiating layer radiates from for each covariate}

\item{smoothing_function}{list(function(n ,mu, covariance, centers)) The function to determine the covariate matrix for the multivariate normal to draw from for each covariate}

\item{radiation_function}{function(x, mu) The function that determines how quickly the radiation decays in the radiating layer}

\item{radiating_means}{list(numeric()) A list with one element for each covariate.  The list elements should have a number for each radiating_polygon for that covariate which is the value that radiates away from that polygon}

\item{weights}{list(numeric(5)) How heavily each layer contributes to the final result for each covariate}

\item{magnitude}{numeric(ncovariates) The variance over gridcells for each covariate (i.e., sets the relative scale for that covariate)}

\item{family}{character Currently unused}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
create multiple test covariates
}
