% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{create_raster_from_shapefile}
\alias{create_raster_from_shapefile}
\title{create_raster_from_shapefile}
\usage{
create_raster_from_shapefile(
  shapefile,
  region = NULL,
  layers_dir = "Layers",
  resol = 10,
  trim = TRUE,
  partial_cover = FALSE,
  trim_small = 0,
  method = "not raster"
)
}
\arguments{
\item{shapefile}{Either a location name, country name, or shapefile to use for the extent of the raster}

\item{region}{A worldpop region containing the shapefile (or NULL to lookup automatically)}

\item{resol}{Approximate resolution of the output raster in km}

\item{trim}{trim raster to shapefile (not just the extent)}

\item{partial_cover}{only used if trim==TRUE, passed to extract_country_from_raster.}

\item{trim_small}{only used if trim==TRUE, passed to extract_country_from_raster.}

\item{method}{only used if trim==TRUE, passed to extract_country_from_raster.}
}
\value{
A RasterLayer
}
\description{
Create a RasterLayer that covers the same region as a shapefile
}
