% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{create_test_covariate}
\alias{create_test_covariate}
\title{create test covariate}
\usage{
create_test_covariate(
  test_raster = create_test_raster(),
  nonspatial = TRUE,
  nontemporal = TRUE,
  spatially_smooth = TRUE,
  temporally_smooth = TRUE,
  polygonal = TRUE,
  radiating = TRUE,
  constant = FALSE,
  rho = 0.999999,
  smoothing_function = function(n, mu, covariance, centers) {    
    return(scale(MASS::mvrnorm(n = n, mu = mu, Matrix::solve(covariance)))) },
  polygons = create_test_layered_polygons(),
  radiation_function = function(x, mu) {     mu * exp(-(x/10000)^2) },
  radiating_polygons = sf::st_sf(sf::st_union(create_test_polygons(dimension = 1))),
  radiating_means = rnorm(nrow(radiating_polygons)),
  weights = c(1, 1, 1, 1, 1),
  family = "Gaussian",
  seed
)
}
\arguments{
\item{test_raster}{raster::raster The raster to make the covariate over.
It should have one layer per time step}

\item{nonspatial}{boolean Whether the independent layer varies in space}

\item{nontemporal}{boolean Whether the independent layer varies in time}

\item{spatially_smooth}{boolean Whether the locally correlated layer is spatially correlated}

\item{temporally_smooth}{boolean Whether the locally correlated layer is temporally correlated}

\item{polygonal}{boolean Whether to include the polygonal layer}

\item{radiating}{boolean Whether to include the radiating layer}

\item{constant}{boolean Whether to include a constant layer}

\item{rho}{numeric A parameter governing spatial correlation used to determine the covariance matrix}

\item{smoothing_function}{function(n ,mu, covariance, centers) The function to determine the covariate matrix for the multivariate normal to draw from}

\item{polygons}{sf::sf The polygons to use for the polygonal layer}

\item{radiation_function}{function(x, mu) The function that determines how quickly the radiation decays in the radiating layer}

\item{radiating_polygons}{sf::sf The polygons that the radiating layer radiates from}

\item{radiating_means}{numeric The values that the radiating layer radiates away from (one for each polygon)}

\item{weights}{numeric(5) How heavily each layer contributes to the final result}

\item{family}{character Currently unused}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Create a test covariate.  The covariate is the sum of various layers. Each layer is governed by various flags and parameters.  The components are:
  independent : each cell is drawn from rnorm(0,1);
  locally correlated : cells are drawn from mvrnorm(0,Sigma), where Sigma is based on the adjacency of gridcells;
  polygonal : Polygons are assigned random values, and each cell is the sum of the value of any polygon it is contained in;
  radiating : Polygons are assigned random values, and each cell is assigned a value based on the distance to each polygon;
  temporal trend : Work in progress;
  and temporal cyclic : Work in progress.
}
