% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{create_test_extent}
\alias{create_test_extent}
\title{create_test_extent}
\usage{
create_test_extent(seed)
}
\arguments{
\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Create an extent to use with modeling
}
