% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_test_layered_polygons}
\alias{create_test_layered_polygons}
\title{create_test_layered_polygons}
\usage{
create_test_layered_polygons(
  test_raster = create_test_raster(),
  base_number = 4,
  n_layers = 3,
  factor = 3,
  snap = FALSE,
  randomize = TRUE,
  seed
)
}
\arguments{
\item{test_raster}{raster::raster The raster to cover with polygons}

\item{base_number}{integer The number of polygons in the coarsest layer}

\item{n_layers}{integer The number of layers of polygons to make.}

\item{factor}{integer The factor by which to increase the number of polygons between layers}

\item{snap}{boolean Whether to snap the edges of the polygon to the boundaries of raster cells}

\item{randomize}{boolean Whether to randomly generate the polygons (as opposed to uniformly)}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Create a layered set of polygons to use for simulating covariates
or observation regions.  Each layer completely covers the same underlying area
}
