% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{create_test_polygons}
\alias{create_test_polygons}
\title{create_test_polygons}
\usage{
create_test_polygons(
  test_raster = create_test_raster(),
  number = 10,
  snap = FALSE,
  randomize = TRUE,
  dimension = 2,
  seed
)
}
\arguments{
\item{test_raster}{raster::raster A raster to cover with interior disjoint polygons}

\item{number}{integer The number of polygons to create}

\item{snap}{boolean Whether to snap the edges of the created polygons to the raster grid}

\item{randomize}{boolean Whether to randomly (as opposed to uniformly) generate the polygons}

\item{dimension}{integer The dimension of polygons to create (0,1, or 2)}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Create a set of polygons to use for simulating covariates or observation regions
}
