% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_test_raster}
\alias{create_test_raster}
\title{create_test_raster}
\usage{
create_test_raster(
  nrows = 20,
  ncols = 20,
  nlayers = 9,
  seed,
  test_extent = create_test_extent(seed)
)
}
\arguments{
\item{nrows}{integer The number of rows the raster should have.}

\item{ncols}{integer The number of columns the raster should have.}

\item{nlayers}{integer The number of layers the raster should have.  Each layer is intended to express time.}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}

\item{test_extent}{The extent the raster should cover.}
}
\description{
Create an raster to use with simulating test data
}
