% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{create_underlying_distribution}
\alias{create_underlying_distribution}
\title{create_underlying_distribution}
\usage{
create_underlying_distribution(
  covariates = create_multiple_test_covariates(),
  normalization = function(x) {     if (length(unique(x[])) == 1) {         return(x *
    0 + 1)     }     x <- exp(x)     x[] <- 0.1 * ((x[] - min(x[]))/(max(x[]) -
    min(x[])))     return(x) },
  family = "Poisson",
  seed
)
}
\arguments{
\item{covariates}{The covariates to use when creating the distribution.  Create using create_test_covariates}

\item{normalization}{The normalization function to use when computing the rate raster.  This function is used to scale the lambda 'covariate' so that it is within a sensible range.}

\item{family}{The family of distribution to use (only Poisson supported for now)}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Create a distribution of ground truth cholera rates
}
