% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{crop_raster_to_shapefile}
\alias{crop_raster_to_shapefile}
\title{crop_raster_to_shapefile}
\usage{
crop_raster_to_shapefile(raster_layer, shapefile)
}
\arguments{
\item{raster_layer}{The Raster* to extract from.}

\item{shapefile}{The shapefile to use for bounding}
}
\value{
A Raster* of the same type as \code{raster_layer} with the extent determined by \code{shapefile}
}
\description{
Crop a raster to the extent of a shapefile
}
