% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{db_exists_table_multi}
\alias{db_exists_table_multi}
\title{Table exists in multiple schemas}
\usage{
db_exists_table_multi(conn, schemas, table_name)
}
\arguments{
\item{conn}{connection to the databas}

\item{schemas}{schemas in which to look for the table}

\item{table_name}{name of the table to look for}
}
\value{
logical
}
\description{
Checks whether a given table exists in multiple schemas
}
