% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{estimate_determinant}
\alias{estimate_determinant}
\title{estimate_determinant}
\usage{
estimate_determinant(rho_samples, nn_mat)
}
\arguments{
\item{rho_samples}{Number of times to sample rho}

\item{nn_mat}{nearest neighbor matrix (see extract_nearest_neighbors_matrix)}
}
\value{
vector of estimates for the determinant of nn_mat
}
\description{
Estimate the determinant of a nearest neighbors matrix
}
