% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_helpers.R
\name{expand_all_locations}
\alias{expand_all_locations}
\title{expand_all_locations}
\usage{
expand_all_locations(.x, .y, assumed_complete_location, all_locations)
}
\arguments{
\item{.x}{dataframe for group_map}

\item{.y}{groups for group_map}

\item{assumed_complete_location}{string with assumed complete location}

\item{all_locations}{list of unique locations}
}
\value{

}
\description{
Helper function to get full list of locations
}
