% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{expand_extent}
\alias{expand_extent}
\title{expand_extent}
\usage{
expand_extent(extent, expansion = 2)
}
\arguments{
\item{extent}{An raster::extent The extent to expand.  Should be the result of raster::expand}

\item{expansion}{A numeric describing how much bigger to make the extent as a numeric.}
}
\description{
Make an extent larger from the center
}
