% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{extract_country_from_raster}
\alias{extract_country_from_raster}
\title{extract_country_from_raster}
\usage{
extract_country_from_raster(
  raster_layer,
  shapefile,
  partial_cover = FALSE,
  method = "raster",
  trim_small = 0,
  tol = 1e-05
)
}
\arguments{
\item{raster_layer}{The Raster* to extract from.}

\item{shapefile}{The shapefile to use for bounding}

\item{partial_cover}{whether or not to do partial matching}

\item{method}{Whether to use the raster method or the previous hard coded one}

\item{trim_small}{Remove values with less than this fraction of the median overlap}

\item{tol}{The distance in (degrees lat/long) rasters are permitted to shift}
}
\value{
A Raster* of the same type as \code{raster_layer}
}
\description{
Extract a Raster* data only for locations in a shapefile
}
