% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{extract_covariate_metadata}
\alias{extract_covariate_metadata}
\title{Extract covariate metadata}
\usage{
extract_covariate_metadata(covar_alias, covar_file, covar_type)
}
\arguments{
\item{covar_alias}{the alias of the covariate to look for in the metadata table}

\item{covar_file}{the file of the raw covariate}

\item{covar_type}{the type of covariate (either 'static' or 'temporal')}
}
\value{
a dataframe with the covariates metadata
}
\description{
Function to extract the metadata from covariates - this could need to be tailored to each covariate
}
