% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{extract_nearest_neighbor_matrix}
\alias{extract_nearest_neighbor_matrix}
\title{extract_nearest_neighbor_matrix}
\usage{
extract_nearest_neighbor_matrix(raster_layer, cell_mask, sparse = T)
}
\arguments{
\item{raster_layer}{The Raster* object to get the cells from}

\item{cell_mask}{The particular cells to extract}

\item{sparse}{Whether to return a sparse matrix or a dense one.  See Matrix for details}
}
\value{
a Matrix object with the nearest neigbhors as 1, and everything else as 0.
}
\description{
Get the nearest neighbor matrix of the chosen cells in the raster
}
