% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{filename_to_stubs}
\alias{filename_to_stubs}
\title{filename_to_stubs}
\usage{
filename_to_stubs(x)
}
\arguments{
\item{x}{filename string or vector of strings}
}
\value{

}
\description{
Parse filename string to get model run information
}
