% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{filter_description_data}
\alias{filter_description_data}
\title{filter_description_data}
\usage{
filter_description_data(data, ...)
}
\arguments{
\item{data}{A data.frame to filter}

\item{...}{As many string arguments as desired.
\itemize{
\item "who_region == 'AFR'"
\item 'ISO_A1 %in% c("COD","NGA")'
\item 'source != "ProMED"'
}
Each filter is applied sequentially, so only data that matches all
  filters will be returned.}
}
\value{
A data.frame with the filters applied
}
\description{
This function selects from a data.frame based on user
  provided filters
}
