% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{fix_a2_l1_name}
\alias{fix_a2_l1_name}
\title{fix_a2_l1_name}
\usage{
fix_a2_l1_name(location_name, country_name = NULL, verbose = TRUE)
}
\arguments{
\item{location_name}{the name of the location to fix}
}
\description{
Change an iso_a2_l1 name to its ISO_3166_1 code
}
