% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookup_helpers.R
\name{fix_country_name}
\alias{fix_country_name}
\title{fix_country_name}
\usage{
fix_country_name(country_name, verbose = TRUE)
}
\arguments{
\item{country_name}{The name of the country}

\item{verbose}{logical}
}
\value{
return ISO country code
}
\description{
Change a country name to its ISO_3166_1 alpha-3 code
}
