% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{gdal_cmd_builder2}
\alias{gdal_cmd_builder2}
\title{gdal_cmd_builder2}
\usage{
gdal_cmd_builder2(
  executable,
  parameter_variables = c(),
  parameter_values = c(),
  parameter_order = c(),
  parameter_noflags = c(),
  parameter_doubledash = c(),
  parameter_noquotes = c(),
  gdal_installation_id = 1,
  python_util = FALSE,
  verbose = FALSE
)
}
\description{
Modified version of gdalUtils::gdal_cmd)builder to allow for unquoted dataset names, which enables the use of postgis layers, the rest of the codes were untouched
}
