% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{gdalinfo2}
\alias{gdalinfo2}
\title{gdalinfo2}
\usage{
gdalinfo2(
  datasetname,
  json,
  mm,
  stats,
  approx_stats,
  hist,
  nogcp,
  nomd,
  nrat,
  noct,
  nofl,
  checksum,
  proj4,
  oo,
  mdd,
  sd,
  version,
  formats,
  format,
  optfile,
  config,
  debug,
  raw_output = TRUE,
  ignore.full_scan = TRUE,
  verbose = FALSE
)
}
\description{
Modified version of gdalUtils::gdalinfo to allow for unquoted dataset names, which enables the use of postgis layers, the rest of the codes were untouched
}
