% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{gdalwarp2}
\alias{gdalwarp2}
\title{gdalwarp2}
\usage{
gdalwarp2(
  srcfile,
  dstfile,
  s_srs,
  t_srs,
  to,
  order,
  tps,
  rpc,
  geoloc,
  et,
  refine_gcps,
  te,
  te_srs,
  tr,
  tap,
  ts,
  ovr,
  wo,
  ot,
  wt,
  r,
  srcnodata,
  dstnodata,
  dstalpha,
  wm,
  multi,
  q,
  of = "GTiff",
  co,
  cutline,
  cl,
  cwhere,
  csql,
  cblend,
  crop_to_cutline,
  overwrite,
  nomd,
  cvmd,
  setci,
  oo,
  doo,
  output_raster = FALSE,
  ignore.full_scan = TRUE,
  verbose = FALSE,
  ...
)
}
\description{
Modified version of gdalUtils::gdalwarp to allow for unquoted dataset names, which enables the use of postgis layers, the rest of the codes were untouched
}
