% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{generate_time_sequence}
\alias{generate_time_sequence}
\title{Generate time sequence}
\usage{
generate_time_sequence(dates, res_time_source, res_time)
}
\arguments{
\item{dates}{Covariate dates}

\item{res_time_source}{Time resolution in units of required time resolution}

\item{res_time}{Time resolution string}
}
\value{
A vector with the sequence of dates
}
\description{
Generates the sequence of times for the modeling covariate in case
the time resolution of the covariate  is coarser than the one required by the model
}
