% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_case_raster}
\alias{get_case_raster}
\title{get_case_raster}
\usage{
get_case_raster(
  preprocessed_data_filename,
  covar_data_filename,
  model_output_filenames
)
}
\arguments{
\item{preprocessed_data_filename}{prepreprocess rdata file name}

\item{covar_data_filename}{covariates rdata filename}

\item{model_output_filenames}{model output filenames}
}
\value{

}
\description{
add
}
