% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_data_fidelity}
\alias{get_data_fidelity}
\title{get_data_fidelity}
\usage{
get_data_fidelity(model_output_filenames)
}
\arguments{
\item{model_output_filenames}{model_output_filenames}
}
\value{

}
\description{
add
}
