% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_filenames}
\alias{get_filenames}
\title{get_filenames}
\usage{
get_filenames(config, cholera_directory)
}
\arguments{
\item{config}{object representing the imported YAML config file for the model}

\item{cholera_directory}{path to cholera directory}
}
\value{

}
\description{
add
}
