% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{get_ncdf_metadata}
\alias{get_ncdf_metadata}
\title{Get NetCDF metadata}
\usage{
get_ncdf_metadata(src_file)
}
\arguments{
\item{src_file}{netcdf file to process}
}
\value{
a list(time_info, var_name, var_att, dates, date_index) with the output
of \code{ncdf4::nc_info} on the time dimension, the variable name, its attributes, the dates
covered by the file and the index they correspond to based on the time dimension
attributes
}
\description{
Gets the metadata on the temporal resolution and coverage of a ncdf file.
}
\details{
The function assumes that the temporal dimension is called 'time'
}
