% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_non_na_gridcells}
\alias{get_non_na_gridcells}
\title{get_non_na_gridcells}
\usage{
get_non_na_gridcells(covar_data_filename)
}
\arguments{
\item{covar_data_filename}{covariates rdata filename}
}
\value{

}
\description{
add
}
