% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_obs_stats}
\alias{get_obs_stats}
\title{get_obs_stats}
\usage{
get_obs_stats(df)
}
\arguments{
\item{df}{sf_cases dataframe with observation data}
}
\value{
dataframe with summary statistics
}
\description{
Prepare observation statistics from the preprocessed rdata object
}
