% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_observed_cells}
\alias{get_observed_cells}
\title{get_observed_cells}
\usage{
get_observed_cells(
  locs,
  pop,
  taxonomy_dir = "taxonomy_verified",
  verbose = FALSE
)
}
\arguments{
\item{locs}{A character vector of location names.  Must be in the standardized form.}

\item{pop}{A raster of population.  Used both to determine where the grids are, and ignore observations with no population}
}
\value{
An integer vector of indices of pop, where the data is included in at least one location and also has non NA population
}
\description{
Get the indices of which cells contain meaningful data given a set of observed locations and a population raster of appropriate size
}
