% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_report_type}
\alias{get_report_type}
\title{get_report_type}
\usage{
get_report_type(taxonomy)
}
\arguments{
\item{taxonomy}{A taxonomy read with read_taxonomy_data}
}
\description{
This function takes all reports in a taxonomy and classifies them by their duration
}
