% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_input_helpers.R
\name{get_space_time_ind_speedup}
\alias{get_space_time_ind_speedup}
\title{Get space time index}
\usage{
get_space_time_ind_speedup(
  df,
  lp_dict,
  model_time_slices,
  res_time,
  do_parallel = T,
  n_cpus = parallel::detectCores() - 2
)
}
\arguments{
\item{df}{the dataframe of the cholera cases}

\item{lp_dict}{the location-periods dictionary}

\item{model_time_slices}{dataframe of left and right bounds of the modeling time slices}

\item{do_parallel}{whether to compute indices in parallel or not}

\item{n_cpus}{the number of cpus to do parellel computatoin}
}
\value{
a list with the mapping to observations, the mapping to grid cells
 and the fraction of modelling time step covered by each observation
}
\description{
Compute the space-time index of observations based on a reference grid and its location-period and time
This version of the code is to run the speedup Stan code which maps observations to location periods and 
location periods to grid cells
}
\details{
the space-time indices are a replication of the grid indices with an offset of the number of pixels.
In this version we compute to mappings: from observations to location periods and from
location periods to grid cells
}
