% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stan_helpers.R
\name{get_stan_parameters}
\alias{get_stan_parameters}
\title{Get stan parameters}
\usage{
get_stan_parameters(config, sigma_eta_scale = 5, beta_sigma_scale = 10)
}
\arguments{
\item{config}{the configuration yaml}

\item{sigma_eta_scale}{the default value for the scale of the yearly random effect}

\item{beta_sigma_scale}{the default value for the scale of the standard deviation of the prior of the regression coefficients}
}
\value{
a list with parameter values
}
\description{
Gets or sets parameters of the stan model
}
