% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{get_temporal_bands}
\alias{get_temporal_bands}
\title{Get temporal bands}
\usage{
get_temporal_bands(model_time_slices, covar_TL_seq, covar_TR_seq)
}
\arguments{
\item{model_time_slices}{dataframe of model time slices (TL and TR)}

\item{covar_TL_seq}{sequence of left time bounds of covariate}

\item{covar_TR_seq}{sequence of right time bounds of covariate}
}
\value{
a list with the band indices, and the left and right time bounds
}
\description{
Get the band numbers corresponding of a temporal covariate
corresponding to the modelling time window
}
