% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{get_time_res}
\alias{get_time_res}
\title{Get time resolution}
\usage{
get_time_res(dates, covar_res_time, units)
}
\arguments{
\item{dates}{Covariate dates}

\item{covar_res_time}{the time resolution of the covariate}

\item{units}{Time resolution units required for the model, one of days, months, years}
}
\value{
Time resolution as numeric rounded to the first digit
}
\description{
Get the time resolution of covariate in required time units
}
