% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{ingest_covariate}
\alias{ingest_covariate}
\title{Ingest covariate}
\usage{
ingest_covariate(
  conn,
  covar_name,
  covar_alias,
  covar_dir,
  covar_unit,
  covar_type,
  covar_res_time = NULL,
  covar_schema,
  ref_grid,
  aoi_extent = NULL,
  aoi_name = "raw",
  res_time,
  time_aggregator = "sum",
  res_x,
  res_y,
  space_aggregator = "mean",
  transform = NULL,
  path_to_cholera_covariates,
  write_to_db = F,
  do_parallel = F,
  n_cpus = 0,
  dbuser
)
}
\arguments{
\item{conn}{connection to the database, object of type DBI::dbConnect}

\item{covar_name}{name of the covariate}

\item{covar_alias}{alias of the covariate in the database}

\item{covar_dir}{directory where to find the covariate}

\item{covar_unit}{not sure}

\item{covar_type}{type of covariate, either 'temporal' or 'static'}

\item{covar_res_time}{default is NULL}

\item{covar_schema}{schema where to save the covariate}

\item{ref_grid}{name of the reference grid in the database}

\item{aoi_extent}{extent of the area of interest to process, default is NULL}

\item{res_time}{string with the input temporal resolution}

\item{time_aggregator}{string with the aggregator to use along the temporal
dimention}

\item{res_x}{longitudinal spatial resolution in km}

\item{res_y}{latitudinal spatial resolution in km}

\item{space_aggregator}{string with the spatial aggregator to use}

\item{transform}{default is NULL}

\item{path_to_cholera_covariates}{path to the trunk of cholera-covariates repository}

\item{write_to_db}{flag to write tables to database or not}

\item{do_parallel}{flag to perform map computations in parallel}

\item{n_cpus}{number of CPUS to run}

\item{dbuser}{User for database connections}
}
\value{
None
}
\description{
Processes a raster covariate at the desired temporal and spatial
resolution and adds it to the database.
}
\details{
The covariates are assumed to be stored in separate folders, each with
potentially multiple multi-band rasters. Covariates of type 'static' will
not be processed temporally
}
