% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_manipulation.R
\name{is_shapefile_valid}
\alias{is_shapefile_valid}
\title{is_shapeilfe_valid}
\usage{
is_shapefile_valid(shapefiles, location_column_name = "id", quiet = FALSE)
}
\arguments{
\item{location_column_name}{The column containing the unique id for the}

\item{shapefiles_to_check}{An sf object}
}
\value{
a data frame with one row for each row of \code{shapefiles_to_check}
  and 4 columns \code{id} The id from the original shapefile \code{valid} is
  a boolean containing whether the shapefile in this position is valid,
  \code{reason} is an explanation for why the shapefile in question is not
  valid, and \code{fixed_value} an attempt to fix the shapefile in question.
}
\description{
Determine whether shapefiles are valid for use in our modeling
  pipeline
}
