% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelists.R
\name{linelist2phantom}
\alias{linelist2phantom}
\title{linelist2phantom}
\usage{
linelist2phantom(
  linelist,
  assumed_complete_location,
  assumed_complete_TL,
  assumed_complete_TR,
  location_columns = "Location",
  case_columns = c("sCh", "cCh", "deaths")
)
}
\arguments{
\item{linelist}{A linelist.  The following columns are required:
- location
- TL 
- TR
- at least one of 
  - suspected_cases
  - confirmed_cases}

\item{assumed_complete_location}{The location the data is assumed complete below}

\item{assumed_complete_TL}{The starting time for the linelist data (defaults to the first time reported)}

\item{assumed_complete_TR}{The ending time for the linelist data (defaults to the first time reported)}

\item{location_columns}{One or more columns containing locations.  If one column is given, delimit location levels by '::'}

\item{case_columns}{One or more columns containing case values to accumulate}
}
\value{
A data_frame with the following columns
 - TL
 - TR
 - case_column
 - each of the location_columns
 - each of the confirmed_cases
}
\description{
Takes a linelist and adds phantom observations where appropriate
}
