% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_orig_map.R
\name{load_orig_map}
\alias{load_orig_map}
\title{load_orig_map}
\usage{
load_orig_map(filename)
}
\arguments{
\item{filename}{object representing the imported YAML config file for the model}
}
\value{
raster of mean or samples of cholera data estimates
}
\description{
load original cholera data estimates (tif or grd) as raster
}
