% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookup_helpers.R
\name{lookup_WorldPop_region}
\alias{lookup_WorldPop_region}
\title{lookup_WorldPop_region}
\usage{
lookup_WorldPop_region(location, verbose = TRUE)
}
\arguments{
\item{location}{The name of the country}

\item{verbose}{Whether to print detailed error messages}
}
\value{
return string of WorldPop region code
}
\description{
Determine which WorldPop raster to look at for population of a country
}
