% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{make_covar_alias}
\alias{make_covar_alias}
\title{Make covariate alias}
\usage{
make_covar_alias(alias, type, res_time, res_space)
}
\arguments{
\item{alias}{covariate alias as specified in the dictionary}

\item{type}{covariate type, either 'temporal' or 'static'}

\item{res_time}{temporal resolution}

\item{res_space}{spatial resolution, vector with two elements for latitude and longitude}
}
\value{
a string with the alias
}
\description{
Function makes the alias of for a given covariate
}
