% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{make_covar_filename}
\alias{make_covar_filename}
\title{Make covariate filename}
\usage{
make_covar_filename(cholera_directory, map_name, covariate_name_part)
}
\arguments{
\item{cholera_directory}{cholera mapping directory}

\item{map_name}{map name}

\item{covariate_name_part}{name of covariate}
}
\value{
a string with the covariate file name
}
\description{
Make string for covariate Rdata file name
}
