% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{make_grid_centroids_table_name}
\alias{make_grid_centroids_table_name}
\title{make grid centroids table name}
\usage{
make_grid_centroids_table_name(dbuser, map_name)
}
\arguments{
\item{dbuser}{database username}

\item{map_name}{A string representing a somewhat unique name for this run}
}
\value{
the table name
}
\description{
makes the table name for the centroids of the computation grid
corresponding to the requested location periods
}
