% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{make_grid_file}
\alias{make_grid_file}
\title{make grid file}
\usage{
make_grid_file(dbuser)
}
\arguments{
\item{dbuser}{database username}
}
\value{
the table name
}
\description{
makes the table name for the grid file
}
