% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{make_locationperiods_table_name}
\alias{make_locationperiods_table_name}
\title{make location periods table name}
\usage{
make_locationperiods_table_name(dbuser, map_name)
}
\arguments{
\item{dbuser}{database username}

\item{map_name}{A string representing a somewhat unique name for this run}
}
\value{
the table name
}
\description{
makes the table name for the requested location periods
}
