% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{make_map_name}
\alias{make_map_name}
\title{Make map name}
\usage{
make_map_name(config, .f = NULL)
}
\arguments{
\item{config}{the configuration file}

\item{.f}{other functions to apply to the config to append to map name}
}
\description{
Make string for map name used for all filenames
}
