% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{make_map_output_filename}
\alias{make_map_output_filename}
\title{Make map output filename}
\usage{
make_map_output_filename(
  cholera_directory,
  map_name,
  covariate_name_part,
  stan_model,
  niter
)
}
\arguments{
\item{cholera_directory}{cholera mapping directory}

\item{map_name}{map name}

\item{covariate_name_part}{name of covariate}

\item{stan_model}{name of stan model}

\item{niter}{number of iterations}
}
\value{
a string with the map output file name
}
\description{
Make string for map pdf file name
}
