% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{make_observations_filename}
\alias{make_observations_filename}
\title{Make observations filename}
\usage{
make_observations_filename(cholera_directory, map_name)
}
\arguments{
\item{cholera_directory}{cholera mapping directory}

\item{map_name}{map name}
}
\value{
a string with the observation file name
}
\description{
Make string for observations Rdata file name
}
