% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{make_stan_input_filename}
\alias{make_stan_input_filename}
\title{Make Stan input filename}
\usage{
make_stan_input_filename(
  cholera_directory,
  map_name,
  covariate_name_part,
  config,
  config_dict
)
}
\arguments{
\item{cholera_directory}{cholera mapping directory}

\item{map_name}{map name}

\item{covariate_name_part}{name of covariate}

\item{config}{config object (not path)}

\item{config_dict}{dictionary object (not path) with abbreviationas of config}
}
\value{
a string with the Stan input file name
}
\description{
Make string for Stan input Rdata file name
}
