% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_functions.R
\name{make_stan_output_filename}
\alias{make_stan_output_filename}
\title{Make Stan output filename}
\usage{
make_stan_output_filename(
  cholera_directory,
  map_name,
  covariate_name_part,
  config,
  config_dict
)
}
\arguments{
\item{cholera_directory}{cholera mapping directory}

\item{map_name}{map name}

\item{covariate_name_part}{name of covariate}

\item{config}{configuration file}

\item{config_dict}{dictionary of configuration options}
}
\value{
a string with the Stan output file name
}
\description{
Make string for Stan output Rdata file name
}
