% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{modeling_time_slices}
\alias{modeling_time_slices}
\title{Modeling time slices}
\usage{
modeling_time_slices(
  start_time,
  end_time,
  res_time,
  time_change_func,
  aggregate_to_start,
  aggregate_to_end
)
}
\arguments{
\item{start_time}{modeling start date (in date format)}

\item{end_time}{modeling end date (in date format)}

\item{res_time}{modeling time resolution}

\item{time_change_func}{function to get the time unit of the date}

\item{aggregate_to_start}{function to get the first date of a time slice}

\item{aggregate_to_end}{function to get the last date of a time slices}
}
\value{
a dataframe left and right bounds of the modeling time slices
}
\description{
Defines the modeling time slicez
}
