% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_helpers.R
\name{na_smart_sum}
\alias{na_smart_sum}
\title{na_smart_sum}
\usage{
na_smart_sum(x)
}
\arguments{
\item{x}{vector to sum}
}
\value{

}
\description{
Helper function to sum vectors with NA values correctly
}
