% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_input_helpers.R
\name{nb2graph}
\alias{nb2graph}
\title{nb2graph}
\usage{
nb2graph(x)
}
\arguments{
\item{x}{nb_object}
}
\value{
dataframe containing num nodes, num edges, and a list of graph edges from node1 to node2. from: https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
\description{
nb2graph
}
