% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{observe_gridcells}
\alias{observe_gridcells}
\title{observe_gridcells}
\usage{
observe_gridcells(
  underlying_distribution = create_underlying_distribution(),
  proportion_observed = 1,
  number_draws = 1,
  spatial_observation_bias = TRUE,
  temporal_observation_bias = TRUE,
  value_observation_bias = TRUE,
  noise = FALSE,
  seed
)
}
\arguments{
\item{underlying_distribution}{The distribution to observe.  Create with create_underlying_distribution}

\item{proportion_observed}{The percent of gridcells whose value is observed}

\item{number_draws}{The number of independent observations to simulate}

\item{spatial_observation_bias}{Whether to choose gridcells to observe in a spatially biased way}

\item{temporal_observation_bias}{Whether to choose gridcells to observe in a temporally biased way}

\item{value_observation_bias}{Whether to choose gridcells to observe biased by the number of cases}

\item{noise}{Whether to add noise to the observations}

\item{seed}{integer A seed to use for the randomly constructed portions of this object}
}
\description{
Simulate observation of the integral of the underlying distribution over the grid
}
