% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework_helpers.R
\name{observe_polygons}
\alias{observe_polygons}
\title{observe_polygons}
\usage{
observe_polygons(
  test_polygons = create_test_layered_polygons(),
  test_covariates = create_multiple_test_covariates(polygons = test_polygons),
  underlying_distribution = create_underlying_distribution(covariates =
    test_covariates),
  grid_proportion_observed = 0.9,
  number_draws = 1,
  grid_spatial_observation_bias = TRUE,
  grid_temporal_observation_bias = TRUE,
  grid_value_observation_bias = TRUE,
  noise = TRUE,
  polygon_proportion_observed = 0.9,
  polygon_observation_rates = exp(rnorm(nrow(test_polygons), -1)),
  polygon_observation_idx = NA,
  polygon_size_bias = TRUE,
  nonlinear_covariates = FALSE,
  min_time_left = lubridate::ymd("2000-01-01"),
  max_time_right = lubridate::ymd("2000-01-01"),
  seed
)
}
\arguments{
\item{test_polygons}{A set of polygons to (potentially) observe cases over}

\item{test_covariates}{A set of covariates to create a distribution out of (not needed if underlying distribution is specified)}

\item{underlying_distribution}{An underlying distribution of cases}

\item{grid_proportion_observed}{What percentage of gridcells get observed by this process}

\item{number_draws}{Number of independent draws from the underlying distribution present}

\item{grid_spatial_observation_bias}{Whether grid observation is correlated to number of cases (see observe_gridcells)}

\item{noise}{Whether or not to include observation noise at the gridcell level}

\item{polygon_proportion_observed}{Percent of polygons observed (for each draw)}

\item{polygon_observation_rates}{Not used}

\item{polygon_size_bias}{Whether or not to sample larger polygons more frequently}

\item{nonlinear_covariates}{Not used}

\item{min_time_left}{The first time associated with an observation}

\item{max_time_right}{The last time associated with an observation}

\item{seed}{A random seed}
}
\description{
observe_polygons
}
