% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{parse_gdal_res}
\alias{parse_gdal_res}
\title{Parse GDAL resolution}
\usage{
parse_gdal_res(src_file)
}
\arguments{
\item{src_file}{raster file to process}
}
\value{
a two-element vector with the longitude and latitude resolutions (in the raster's coordinate system)
}
\description{
Parses the output of gdalinfo to extract the pixel resolution
of a raster
}
