% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{parse_time_res}
\alias{parse_time_res}
\title{Parse time resolution}
\usage{
parse_time_res(x)
}
\arguments{
\item{x}{string of the time resolution}
}
\value{
a list(units, k) with the units and the number of single units
the temporal resolution corresponds to.
}
\description{
Function to parse the string of the time step of the model into
the time unit and the spacing for the \code{rts} package
}
\examples{
parseTimeRes("1 month")
# $units
# [1] "months"
# $k
# [1] 1
}
