% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_area_adjusted_observed_cases}
\alias{plot_area_adjusted_observed_cases}
\alias{plot_raw_observations}
\title{plot_area_adjusted_observed_cases}
\usage{
plot_area_adjusted_observed_cases(
  disjoint_set_sf_cases,
  render = F,
  plot_file = NULL,
  width = NULL,
  height = NULL
)

plot_raw_observations(
  disjoint_set_sf_cases,
  render = F,
  plot_file = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{disjoint_set_sf_cases}{disjoint set of sf cases object}

\item{render}{default is FALSE}

\item{plot_file}{default is NULL}

\item{width}{plot width}

\item{height}{plot height}
}
\value{
ggplot object with area-adjusted observed cases

ggplot object with number of observations observed by unique location periods
}
\description{
add

add
}
