% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_chain_convergence}
\alias{plot_chain_convergence}
\title{plot_chain_convergence}
\usage{
plot_chain_convergence(
  model_output_filenames,
  pars = c("rho", "betas", "log_std_dev_w"),
  render = T
)
}
\arguments{
\item{model_output_filenames}{model output filenames}

\item{pars}{parameters for which to display traceplots}

\item{render}{default is TRUE}
}
\value{
ggplot object with traceplots by parameter
}
\description{
add
}
