% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(
  map_output,
  column = "cases",
  facet_column = "type",
  colorscale_type = "cases",
  render = FALSE,
  plot_file = NULL,
  width = NULL,
  height = NULL,
  plot_border = TRUE,
  ...
)
}
\arguments{
\item{map_output}{sf or raster object containing data to plot}

\item{column}{string with column name to plot, typically `cases` or `rate` (default: `cases`)}

\item{facet_column}{string with column name to facet}

\item{colorscale_type}{string with type of default color scale to use (default: `cases`) (See [`color_scale()`] for more details)}

\item{render}{logical for whether to render figure (default: FALSE)}

\item{plot_file}{if NULL, save to file (default: NULL)}

\item{width}{width of saved figure in inches}

\item{height}{height of saved figure in inches}

\item{plot_border}{logical to include grid borders}

\item{...}{additional parameters passed to [`color_scale()`]}
}
\value{
ggplot object of cholera case or incidence maps on standard color scale
}
\description{
default function to plot cholera maps with ggplot
}
