% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_model_fidelity}
\alias{plot_model_fidelity}
\title{plot_model_fidelity}
\usage{
plot_model_fidelity(data_fidelity, case_raster, render = T)
}
\arguments{
\item{data_fidelity}{data_fidelity object}

\item{case_raster}{case_raster object}

\item{render}{default is TRUE}
}
\value{
ggplot object with modeled vs actual cases by observation
}
\description{
add
}
