% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_modeled_rates}
\alias{plot_modeled_rates}
\title{plot_modeled_rates}
\usage{
plot_modeled_rates(
  case_raster,
  render = T,
  plot_file = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{case_raster}{case_raster object}

\item{render}{default is TRUE}

\item{plot_file}{default is NULL}

\item{width}{plot width}

\item{height}{plot height}
}
\value{
ggplot object with modeled rates map
}
\description{
add
}
