% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_raster_covariates}
\alias{plot_raster_covariates}
\title{plot_raster_covariates}
\usage{
plot_raster_covariates(covar_data_filename, render = T)
}
\arguments{
\item{covar_data_filename}{covariates rdata filename}

\item{render}{default is FALSE}
}
\value{
ggplot object with covariate raster
}
\description{
add
}
