% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{plot_rhat}
\alias{plot_rhat}
\title{plot_rhat}
\usage{
plot_rhat(model.rand, render = T)
}
\arguments{
\item{model.rand}{Stan model output}

\item{render}{default is TRUE}
}
\value{
ggplot object with Rhat by observation id
}
\description{
add
}
