% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{pull_output_by_source}
\alias{pull_output_by_source}
\title{pull_output_by_source}
\usage{
pull_output_by_source(
  sf_cases,
  source_match,
  database_api_key_rfile = "../R/database_api_key.R"
)
}
\arguments{
\item{sf_cases}{sf_cases output}

\item{source_match}{string of reference source (e.g., WHO Annual Cholera Reports)}
}
\value{

}
\description{
Most of this function is a hack to get the source for sql pulls.  The real fix for this problem is to have the sql pull already have the source.
}
