% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{pull_taxonomy_data}
\alias{pull_taxonomy_data}
\title{Pull taxonomy data}
\usage{
pull_taxonomy_data(
  username,
  password,
  locations = NULL,
  time_left = NULL,
  time_right = NULL,
  uids = NULL,
  website = "https://api.cholera-taxonomy.middle-distance.com/",
  source
)
}
\arguments{
\item{username}{taxonomy username}

\item{password}{taxonomy password}

\item{locations}{list of locations to pull. For now this only supports country ISO codes.}

\item{time_left}{left bound for observation times (in date format)}

\item{time_right}{right bound for observation times (in date format)}

\item{uids}{list of unique observation collection ids to pull}

\item{website}{Which website to pull from (default is cholera-taxonomy.middle-distance.com)}

\item{source}{whether to pull data from the website or using sql on idmodeling2. 
Needs to be one of 'api' or 'sql'.}

\item{api_key}{A working api.key for the user of the database}
}
\value{
An sf object containing data pulled from the database
}
\description{
Pulls data from the taxonomy database
}
\details{
This is a wrapper which calls either read_taxonomy_data_database or
read_taxonomy_data_sql depending on the source that the user specifies.
}
