% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{rdist_stan_model}
\alias{rdist_stan_model}
\title{rdist_stan_model}
\usage{
rdist_stan_model(n, result)
}
\arguments{
\item{n}{The number of times to sample}

\item{result}{The results of a stan model to draw from}
}
\description{
rdist_stan_model
}
