% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_description_taxonomy}
\alias{read_description_taxonomy}
\title{read_description_taxonomy}
\usage{
read_description_taxonomy(taxonomy.directory, ..., uids)
}
\arguments{
\item{taxonomy.directory}{A string for the path for the directory
described above.}

\item{...}{A sequence of filters used to filter the description
files.  See \code{filter_description_data} for details}

\item{uids}{A vector of uids to read the description files for.
This parameter is optional.  If missing, this function will read
all present uids.}
}
\value{
A data.frame containing the data read from the directory
  \code{taxonomy.directory} filtered by the filters \code{...}
}
\description{
This function will pull all of the description
  information.  It reads all of the description files, and combines
  them into a single data.frame
}
