% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_epi_taxonomy}
\alias{read_epi_taxonomy}
\title{read_epi_taxonomy}
\usage{
read_epi_taxonomy(taxonomy.directory, uids, verbose = FALSE)
}
\arguments{
\item{taxonomy.directory}{A string for the path for the directory
described above.}

\item{uids}{A vector of uids to read the description files for.
This parameter is optional.  If missing, this function will read
all present uids.}

\item{columns}{From the final data, which columns to select before
returning the data.}
}
\value{
A data.frame containing the specified data read from the
  directory \code{taxonomy.directory} filtered by the filters
  \code{...}
}
\description{
This function will pull all of the epi case
  information.  It reads all of the epi files, and combines them
  into a single data.frame
}
