% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{read_file_of_type}
\alias{read_file_of_type}
\title{read_file_of_type}
\usage{
read_file_of_type(filename, variable)
}
\arguments{
\item{filename}{filename}

\item{variable}{variable present in file}
}
\value{

}
\description{
Read pipeline files
}
