% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\docType{data}
\name{read_location_csv}
\alias{read_location_csv}
\alias{location_coltypes}
\title{read_location_csv}
\format{
An object of class \code{character} of length 30.
}
\usage{
location_coltypes
}
\arguments{
\item{filename}{A string for the relative or absolute path to the
file to be read.  The file format is described in the
documentation}
}
\value{
A \code{data.frame} containing the data from
  \code{filename} or a single column missing with the filename
  listed
}
\description{
This function reads a csv file containing a location file
}
\keyword{datasets}
