% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\docType{data}
\name{read_population_csv}
\alias{read_population_csv}
\alias{population_coltypes}
\title{read_population_csv}
\format{
An object of class \code{character} of length 4.
}
\usage{
population_coltypes
}
\arguments{
\item{filename}{A string for the relative or absolute path to the
file to be read.  The file format is described in the
documentation}
}
\value{
A \code{data.frame} containing the data from
  \code{filename} or a single column missing with the filename
  listed
}
\description{
This function reads a csv file containing a population
  file
}
\keyword{datasets}
