% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_taxonomy_csv}
\alias{read_taxonomy_csv}
\title{read_taxonomy_csv}
\usage{
read_taxonomy_csv(filename, verbose = FALSE)
}
\arguments{
\item{filename}{A string for the relative or absolute path to the
file to be read.  The file should be column major}
}
\value{
A data.frame containing the data from \code{filename} or a
  data.frame with a single column missing=\code{filename})
}
\description{
This function reads a csv file which is column major.
  It will also fail without throwing an error, so it can be used on
  missing data
}
\details{
Taxonomy Data Parser

The Taxonomy Data Parser package loads and handles Taxonomy Data.
It provides 1 category of functions:
  Taxonomy Parsing

Taxonomy parsing functions are designed to read flat
  files from a filesystem storing taxonomy data.  If the file
  system is called taxonomy.directory, then the following
  subdirectories must exist:
\itemize{
\item taxonomy.directory/Description - files must end in _DESC.csv
\item taxonomy.directory/Location    - files must end in _LOC.csv
\item taxonomy.directory/EpiCurves   - files must end in _EPI.csv
\item taxonomy.directory/Population  - files must end in _POP.csv
}
The files Description and Population files are regular .csv files,
  and the Location and Epi files are transposed .csv files.
}
\section{Taxonomy Parsing Functions}{

}

