% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_taxonomy_data}
\alias{read_taxonomy_data}
\title{read_taxonomy_data}
\usage{
read_taxonomy_data(
  taxonomy.directory = "taxonomy-verified",
  columns = NULL,
  ...
)
}
\arguments{
\item{taxonomy.directory}{A string for the path for the directory
described above.}

\item{columns}{From the final data, which columns to select before
returning the data.}

\item{...}{A sequence of filters used to filter the description
files.  See \code{filter_description_data} for details}
}
\value{
A data.frame containing the specified \code{columns} of the
  data read from the directory \code{taxonomy.directory} filtered
  by the filters \code{...}
}
\description{
This function is the main function in this package.
  It reads the taxonomy files from the filesystem, and combines
  them all into a single data.frame
}
