% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_helpers.R
\name{read_taxonomy_data_api}
\alias{read_taxonomy_data_api}
\title{read_taxonomy_data_api}
\usage{
read_taxonomy_data_api(
  username,
  api_key,
  locations = NULL,
  time_left = NULL,
  time_right = NULL,
  uids = NULL,
  website = "https://api.cholera-taxonomy.middle-distance.com/"
)
}
\arguments{
\item{username}{The username for a user of the database}

\item{api_key}{A working api.key for the user of the database}

\item{locations}{A vector of locations to pull observations from (should be in the form who_region::ISO_L1::ISO_A2_...)}

\item{time_left}{First time for observations}

\item{time_right}{Last time for observations}

\item{uids}{unique observation collections ids to pull}

\item{website}{Which website to pull from (default is cholera-taxonomy.middle-distance.com)}
}
\value{
An sf object containing data pulled from the database
}
\description{
This function accesses the cholera-taxonomy stored
  at https://api.cholera-taxonomy.middle-distance.com pulls
  data based on function parameters, links it together, and
  transforms it into a simple features object (sf).
}
