% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_transposed_taxonomy_csv}
\alias{read_transposed_taxonomy_csv}
\title{read_transposed_taxonomy_csv}
\usage{
read_transposed_taxonomy_csv(filename, verbose = FALSE)
}
\arguments{
\item{filename}{A string for the relative or absolute path to the
file to be read.  The file should be row major}
}
\value{
A data.frame containing the data from \code{filename} or a
  data.frame with a single column missing=\code{filename})
}
\description{
This function reads a csv file which is row major.  It
  will also fail without throwing an error, so it can be used on
  missing data
}
