% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{regrid_raster}
\alias{regrid_raster}
\title{regrid_raster}
\usage{
regrid_raster(x, y, aggregate = TRUE, .fun = sum)
}
\arguments{
\item{x}{The Raster* object to regrid}

\item{y}{The Raster* object to obtain the extent,resolution etc from}
}
\value{
A Raster* of the same type and layers as x with the same extent/resolution/etc as y
}
\description{
using linear combinations, put one raster on the grid of another raster
}
