% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_helpers.R
\name{rename_database_fields}
\alias{rename_database_fields}
\title{Rename cholera data columns}
\usage{
rename_database_fields(database_df, source = "api")
}
\arguments{
\item{database_df}{Data who's columns are to be modified}

\item{source}{Whether the source is the staging database (sing the API) or taxdat (using SQL).}
}
\value{
the renamed dataframe
}
\description{
Renames the columns of the data pulled either from the the 
API staging database or by SQL from taxdat
}
\details{
source is one of 'api' or 'sql'
}
