% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{safe.function}
\alias{safe.function}
\title{safe.function}
\usage{
safe.function(fxn, ..., default = warning("The function failed to run"))
}
\arguments{
\item{\dots}{The normal arguments to the function}

\item{default}{The value the function should return if it fails}

\item{function}{The function to run safely}
}
\value{
The return value of the function if the function runs,
  or \code{default} otherwise
}
\description{
This function runs a function, and returns either
  the return of that function, or a default return if it fails
}
