% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{sample_stan_model}
\alias{sample_stan_model}
\title{sample_stan_model}
\usage{
sample_stan_model(n, result, locs, pop_raster)
}
\arguments{
\item{n}{The number of times to sample}

\item{result}{The model results to sample from}

\item{locs}{A character vector of location names where there are observations}

\item{pop_raster}{The underlying population raster.  Used for determining the raster shape and finding places with no population}
}
\description{
get samples from the results of a stan model
}
