% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{show_progress}
\alias{show_progress}
\title{Show progress}
\usage{
show_progress(i, tot_length, n_hyph_tot = 80, prefix = "")
}
\arguments{
\item{i}{current step}

\item{tot_length}{total number of steps to do}

\item{n_hyph_tot}{number of hyphens corresponding to the full completion}

\item{prefix}{string to paste at the beginning of the bar}
}
\value{
None
}
\description{
Shows a bar of the progess of a process based on a scalar iterator
}
