% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{space_aggregate}
\alias{space_aggregate}
\title{Spatial aggregation}
\usage{
space_aggregate(
  src_file,
  res_file,
  ref_grid_db,
  covar_type,
  aggregator,
  dbuser
)
}
\arguments{
\item{src_file}{raster file to process}

\item{res_file}{file to which to write the aggregated raster should be written to}

\item{ref_grid_db}{reference grid}

\item{covar_type}{type of covariate, one of 'temporal' or 'static'}

\item{aggregator}{string defining the aggregator to use}

\item{dbuser}{the database user name}
}
\value{
None
}
\description{
Aggregate raster in space based on a reference raster
}
