% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{st_update_bounding_box}
\alias{st_update_bounding_box}
\title{st_update_bounding_box}
\usage{
st_update_bounding_box(shp)
}
\arguments{
\item{shp}{an sf object}
}
\value{
a copy of shp with a modified bounding box
}
\description{
Make sure the bounding box of a sf object is correct
}
