% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{stan_result_apply}
\alias{stan_result_apply}
\alias{stan_result_summarize}
\title{stan_result_apply}
\usage{
stan_result_summarize(result, fun)
}
\arguments{
\item{result}{The result of a (particular kind of) stan model}

\item{fun}{The function to apply to summarize the results}
}
\description{
stan_result_apply
}
