% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookup_helpers.R
\name{standardize_string}
\alias{standardize_string}
\title{standardize_string}
\usage{
standardize_string(string)
}
\arguments{
\item{string}{character string}
}
\value{
standardize string
}
\description{
standardize_string
}
