% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{sum_nonNA}
\alias{sum_nonNA}
\title{Sum non-NA cells}
\usage{
sum_nonNA(conn, r_file, ref_grid_db, dbuser)
}
\arguments{
\item{conn}{the postgres connection to the database}

\item{r_file}{the processed raster file}

\item{ref_grid_db}{the reference grid in the database}

\item{dbuser}{the database user name}
}
\value{
A raster with the number of non-NA cells
}
\description{
Sums the number of non-NA cells in a resampled raster
}
