% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{time_aggregate}
\alias{time_aggregate}
\title{Aggregate time}
\usage{
time_aggregate(
  src_file,
  covar_name,
  covar_unit,
  covar_type,
  covar_res_time,
  res_file,
  res_time,
  aggregator,
  aoi_extent = NULL
)
}
\arguments{
\item{src_file}{raster file to process}

\item{covar_name}{name of the covariate}

\item{covar_unit}{units of the covariate}

\item{covar_type}{type of covariate, one of 'temporal' or 'static'}

\item{covar_res_time}{string indicating time resolution of covariate}

\item{res_file}{file to which to write the aggregated temporal raster}

\item{res_time}{time resolution string}

\item{aggregator}{string defining the aggregator to use}

\item{aoi_extent}{extent of the area of interest to process}
}
\value{
a string with the path to the result
}
\description{
Aggregates a multi-band temporal raster at a desired temporal
resolution
}
\details{
For now the aggregator string needs to be an R function that exists in
the global enviroment.
}
