% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{time_overlap}
\alias{time_overlap}
\title{Time overlap}
\usage{
time_overlap(tl, tr, res_time, tl_vec, tr_vec)
}
\arguments{
\item{tl}{time left, starting time}

\item{tr}{time right, ending time}

\item{tl_vec}{vector of time lefts}

\item{tr_vec}{vector of time rights}
}
\value{
a list with the time overlaps
}
\description{
Computes the time overlap between a date range and a vector of ranges
}
