% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{time_unit_to_aggregate_function}
\alias{time_unit_to_aggregate_function}
\title{time_unit_to_aggregate_function}
\usage{
time_unit_to_aggregate_function(unit)
}
\arguments{
\item{unit}{Human readable unit of time aggregation}
}
\value{
function to convert dates to the right thing
}
\description{
Turns the
}
