% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{transform_raster}
\alias{transform_raster}
\title{Transform raster}
\usage{
transform_raster(res_file, transform)
}
\arguments{
\item{res_file}{the raster to transform}

\item{transform}{the transform to apply}
}
\value{
none (writes resulting rater to file)
}
\description{
transforms a raster with a user-specified function
}
