% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata}
\usage{
write_metadata(
  conn,
  covar_dir,
  covar_type,
  covar_alias,
  res_x,
  res_y,
  res_time,
  space_aggregator,
  time_aggregator,
  dbuser
)
}
\arguments{
\item{conn}{database connection}

\item{covar_dir}{directory where to find the covariate}

\item{covar_alias}{alias of the covariate in the database}

\item{res_x}{longitudinal spatial resolution in km}

\item{res_y}{latitudinal spatial resolution in km}

\item{res_time}{string with the input temporal resolution}

\item{space_aggregator}{string with the spatial aggregator to use}

\item{time_aggregator}{string with the aggregator to use along the temporal
dimention}

\item{dbuser}{user name}
}
\value{
return
}
\description{
writes the covariate metadata to the metadata table
}
