% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{write_ncdf}
\alias{write_ncdf}
\title{Write NCDF}
\usage{
write_ncdf(
  data,
  res_file,
  mv = -9999,
  var_name,
  long_var_name,
  var_unit,
  time_units,
  time_vals,
  chunk_size = 1000
)
}
\arguments{
\item{data}{the raster data to write}

\item{res_file}{the file to which to write the result}

\item{mv}{missing value}

\item{var_name}{the name of the variable}

\item{long_var_name}{the long name of the variable}

\item{var_unit}{the unit of the variable}

\item{time_units}{the units of the time dimension}

\item{time_vals}{the values of the time dimension}
}
\value{
null
}
\description{
writes a netcdf file from a raster object
}
\details{
From https://gis.stackexchange.com/questions/58550/r-raster-package-write-netcdf-with-time-dimension
}
