% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helpers.R
\name{write_pg_raster}
\alias{write_pg_raster}
\title{Write postgis raster}
\usage{
write_pg_raster(dbuser, schema, table, outfile, band = 1)
}
\arguments{
\item{dbuser}{user name}

\item{schema}{schema read the raster from}

\item{table}{raster table name}

\item{outfile}{file to which to write the raster}
}
\description{
Write raster from postgis table
}
\details{
currently only works for 1 band rasters
from https://gis.stackexchange.com/questions/118138/how-to-load-postgis-raster-layers-into-r
}
