% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_helpers.R
\name{get_disjoint_set_sf_cases}
\alias{get_disjoint_set_sf_cases}
\title{get_disjoint_set_sf_cases}
\usage{
get_disjoint_set_sf_cases(
  .x = NULL,
  preprocessed_data_filename = NULL,
  stan_input_filename = NULL
)
}
\arguments{
\item{preprocessed_data_filename}{Rdata filename with the preprocess suffix}
}
\value{

}
\description{
add
}
