% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_helpers.R
\name{linelist2phantom}
\alias{linelist2phantom}
\title{linelist2phantom}
\usage{
linelist2phantom(
  linelist,
  assumed_complete_location,
  assumed_complete_TL,
  assumed_complete_TR,
  location_columns = "Location",
  case_columns = c("sCh", "cCh", "deaths")
)
}
\arguments{
\item{linelist}{A linelist.  The following columns are required:
- location
- TL 
- TR
- at least one of 
  - sCh (suspected cases)
  - cCh (confirmed cases)
  - deaths}

\item{assumed_complete_location}{The location below which the data is assumed complete (e.g., "AFR::ETH" for a linelist representing the whole country of Ethiopia)}

\item{assumed_complete_TL}{The starting time for the linelist data (defaults to the first time reported in the linelist)}

\item{assumed_complete_TR}{The ending time for the linelist data (defaults to the first time reported)}

\item{location_columns}{The names of one or more columns containing locations (e.g., "Location" or c("ISO_A1", "ISO_A2_L1")). If only one column is given, delimit location levels by '::'}

\item{case_columns}{One or more columns containing case values across which to accumulate. Defaults to sCh, cCh, and deaths.}
}
\value{
A data_frame with the following columns
 - TL
 - TR
 - case_column
 - each of the location_columns
 - each of the confirmed_cases
}
\description{
Takes a line ist and adds daily primary phantom observations for locations and dates not present in the linelist
}
