% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_helpers.R
\name{read_taxonomy_data_sql}
\alias{read_taxonomy_data_sql}
\title{Taxonomy SQL data pull}
\usage{
read_taxonomy_data_sql(
  username,
  password,
  locations = NULL,
  time_left = NULL,
  time_right = NULL,
  uids = NULL
)
}
\arguments{
\item{username}{taxonomy username}

\item{password}{taxonomy password}

\item{locations}{list of locations to pull. For now this only supports country ISO codes.}

\item{time_left}{left bound for observation times (in date format)}

\item{time_right}{right bound for observation times (in date format)}

\item{uids}{list of unique observation collection ids to pull}
}
\value{
An sf object containing data extracted from the database
}
\description{
Extracts data for a given set of country using SQL from the taxonomy
postgresql database stored on idmodeling2
}
\details{
Code follows taxdat::read_taxonomy_data_api template.
}
