% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_helpers.R
\name{time_unit_to_aggregate_function}
\alias{time_unit_to_aggregate_function}
\title{time_unit_to_aggregate_function}
\usage{
time_unit_to_aggregate_function(unit)
}
\arguments{
\item{unit}{Human readable string for unit of time aggregation}
}
\value{
a function that converts dates into the correct time unit
}
\description{
Returns a function that converts dates into the correct time unit
}
