% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{time_unit_to_end_function}
\alias{time_unit_to_end_function}
\title{time_unit_to_end_function}
\usage{
time_unit_to_end_function(unit)
}
\arguments{
\item{type}{string of type}
}
\value{
function to convert dates to the right thing
}
\description{
Turns human readable time units into a functions that convert time units to the end of the time unit
}
