% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_helpers.R
\name{time_unit_to_start_function}
\alias{time_unit_to_start_function}
\title{time_unit_to_start_function}
\usage{
time_unit_to_start_function(unit)
}
\arguments{
\item{unit}{Human readable string for unit of time aggregation}
}
\value{
a function that returns a date at the start of the appropriate time unit
}
\description{
Make a function that returns a date at the start of the appropriate time unit
}
